pyproject.toml
[tool.poetry]
name = "spotinoty"
version = "0.1.0"
description = ""
authors = ["altvec <me@sergeykalistratov.com>"]
license = "MIT"
packages = [
    { include = "spotinoty" },
]

[tool.poetry.dependencies]
python = "^3.9"
spotipy = "^2.16.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.14.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
spotinoty = "spotinoty.scripts.spotinoty:main"
